requirements.txt
alembic
email-validator
flask
flask-login
flask-bootstrap
flask-wtf
flask-sqlalchemy
flask-migrate
flask-mail
flask-moment
flask-babel
pyjwt
WTForms
Werkzeug
SQLAlchemy
click
guess-language_spirit
requests
python-dotenv